Update from original (#1)
* Added support for OpenRC in the installer script.

This change has been tested on the following systems,
- alpine:latest
- gentoo:stage3-amd64
- ubuntu:14.04

Signed-off-by: Lucas Ramage <ramage.lucas@openmailbox.org>

* Docs: Update vagrant-spk key-handling

* make update-deps

* meteor update

* Update changelog for v0.239.

* Fix install.sh syntax.

* Unauthenticated API host should serve with access-control-allow-origin: *

This change is necessary for the tests to pass. But somehow the tests historically passed without it, like the browser wasn't checking CORS somehow...

* Modernize test suite.

The ancient version of Selenium we were using no longer works with the version of Firefox ESR on Debian Sid. Newer Seleniums no longer build-in Firefox support but rather require separately-installed drivers. Nightwatch also needed to be updated to work with newer Selenium. I switched to installing Selenium via npm as this seems more convenient.

For Firefox, the driver is `geckodriver`, but Nightwatch has a bug where its switchWindow() function doesn't work with `geckodriver`. The latest beta version of Nightwatch fixes this bug but seems to intorduce some other bugs. Instead, I switched to running the tests in Chrome.

Various tests needed updates to work correctly on the new version.

* Make sure to display note about free plan discontinuation even after the switch.

* Extend 'quota exceeded' error to link to blog post with more details.

* make update-deps

* meteor update

* Update babel/runtime as instructed by error message on console.

* Work around https://github.com/meteor/meteor/issues/10290

* Update changelog for v0.240.

* We have Gateways now.

* Update release certificate.

* meteor update

* meteor npm update --save

* make update-deps

* Update selenium to work with JDK 11.

* Update changelog for v0.241.

* Update release certificate.

* make update-deps

* meteor update

* meteor update --all-packages (then revert find-and-modidy)

* meteor npm update --save

* meteor npm update --depth 9999 --save

* Add newly-required include.

* Update changelog for v0.242.

* npm-shrinkwrap.json updated itself.

* 'delete my account' must be typed in English regardless of language setting.

A French user complained they were unable to delete their account and eventually determined the instructions were wrong.

All the other translations correctly show the quote in English.

* make update-deps

* meteor update

* meteor npm update --save

* npm-shrinkwrap.json updated itself.

* Fix tests/run-local.sh to make sure it solely uses Meteor's node due to apparent incompatibility with newer system-installed node.

* Update changelog for v0.243.

* Try to fix intrinsicNames issue.

* Typo fix in admin panel i18n

Fixes #3115

* shell/npm-shrinkwrap.json updated itself some time in the past?

* Update release certificate.

* make update-deps

* meteor update

* meteor npm update --save

* Implement KJ's new whenWriteDisconnected().

* npm-shrinkwrap.json updated itself

* Update changelog for v0.244.

* make update-deps

* meteor update

* meteor npm update --save

* npm-shrinkwrap.json updated itself

* Update changelog for v0.245.

* make update-deps

* meteor npm update --save

* Update changelog for v0.246.

* npm-shrinkwrap.json updated itself

* Allow Ekam binary location to be specified by make variable.

* make update-deps

* meteor update

* meteor npm update --save

* meteor npm audit fix

Note: All the vulnerabilities found by `npm audit` were in build dependencies, not runtime dependencies. (In fact, they were all build dependencies of the "bignum" package.) So, it does not appear that this ever affected Sandstorm. But, I didn't want to ignore the warnings printed by npm.

* Switch to using Meteor's build of Node rather than our own, since it now includes the patch we needed.

* Update changelog for v0.247.

* make update-deps

* meteor npm update --save

* Update changelog for v0.248

* make update-deps

* meteor update

* meteor npm update --save

* Update for Cap'n Proto API changes.

* Fix a warning.

* Update changelog for v0.249.

* Update capnproto to remove obsolete TODO(now).

* Introduce schemas to be able to restore user identities after grain transfer.

* Remember identity IDs for forgotten grains.

Before this change, if someone shared a grain with you, you opened it, then you deleted it from your grain list, and then later on they shared it with you again, you would get a different identity ID the second time, and so would appear to be a different user within the grain.

After this change, you will regain your original identity ID.

* Remember identities through grain backup/restore.

Previously, when a grain was backed up and then restored, if the restored grain was shared with the same users as the original, the app would not know that they were the same users, as they would be assigned new, different identity IDs. After this change, identities will be remembered as long as the new user shares at least one of the original user's login credentials.

* meteor update

* meteor npm update --save

* make update-deps

* Update changelog for v0.250.

* Implement Mass Transfers.

This makes it easy to mass-transfer grains to another Sandstorm server. A UI lets you choose which grains to transfer, watch the transfer happen, and check and retry any errors.

* make update-deps

* Update changelog for v0.251.

* Fix bug applying only to Blackrock, ugh.

* Update release certificate.

* Fix `spk dev` resulting in "This grain's app package is not installed".

This error was introduced in the Mass Transfers PR from last weekend. I was trying to improve the UI when a grain's package was not installed, but the new checks I added didn't consider when the app is registered via `spk dev`.

* Update changelog for v0.252.

* Typo in Mass Transfer page

s/cat perform/can perform